@@ -21,7 +21,6 @@
     "@ebay/nice-modal-react": "1.2.13",
     "@fontsource/dejavu-sans": "5.0.3",
     "@fontsource/public-sans": "5.0.3",
-    "@hot-loader/react-dom": "17.0.1",
     "@opentrons/api-client": "link:../api-client",
     "@opentrons/components": "link:../components",
     "@opentrons/react-api-client": "link:../react-api-client",
@@ -43,10 +42,9 @@
     "mixpanel-browser": "2.22.1",
     "netmask": "2.0.2",
     "path-to-regexp": "3.0.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-error-boundary": "^4.0.10",
-    "react-hot-loader": "4.13.0",
     "react-i18next": "^11.7.3",
     "react-intersection-observer": "^8.33.1",
     "react-redux": "7.2.1",
