@@ -18,7 +18,7 @@
   },
   "homepage": "https://github.com/Opentrons/opentrons#readme",
   "peerDependencies": {
-    "react": "17.0.1",
+    "react": "18.2.0",
     "react-router-dom": "5.1.1"
   },
   "dependencies": {
@@ -28,7 +28,6 @@
     "@react-spring/web": "9.6.1",
     "@types/classnames": "2.2.5",
     "@types/lodash": "^4.14.191",
-    "@types/react": "^17.0.1",
     "@types/react-router-dom": "^5.1.7",
     "@types/styled-components": "^5.1.26",
     "@types/webpack-env": "^1.16.0",
