@@ -28,7 +28,6 @@
     "@types/yup": "0.29.11"
   },
   "dependencies": {
-    "@hot-loader/react-dom": "17.0.1",
     "@opentrons/components": "link:../components",
     "ajv": "6.12.3",
     "classnames": "2.2.5",
@@ -40,9 +39,8 @@
     "lodash": "4.17.21",
     "mixpanel-browser": "2.29.1",
     "query-string": "6.2.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-hot-loader": "4.13.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-router-dom": "5.1.1",
     "yup": "0.32.9"
   }
