@@ -19,7 +19,6 @@
   "homepage": "https://github.com/Opentrons/opentrons",
   "license": "Apache-2.0",
   "dependencies": {
-    "@hot-loader/react-dom": "17.0.1",
     "@opentrons/components": "link:../components",
     "@opentrons/step-generation": "link:../step-generation",
     "@opentrons/shared-data": "link:../shared-data",
@@ -37,12 +36,11 @@
     "immer": "9.0.6",
     "lodash": "4.17.21",
     "query-string": "6.2.0",
-    "react": "17.0.1",
+    "react": "18.2.0",
     "react-color": "2.19.3",
     "react-dnd": "6.0.0",
     "react-dnd-mouse-backend": "0.1.2",
-    "react-dom": "17.0.1",
-    "react-hot-loader": "4.13.0",
+    "react-dom": "18.2.0",
     "react-redux": "7.2.1",
     "redux": "4.0.5",
     "redux-actions": "2.2.1",
