@@ -20,13 +20,11 @@
   "homepage": "https://github.com/Opentrons/opentrons",
   "license": "Apache-2.0",
   "dependencies": {
-    "@hot-loader/react-dom": "17.0.1",
     "@opentrons/components": "link:../components",
     "@opentrons/shared-data": "link:../shared-data",
     "classnames": "2.2.5",
     "lodash": "4.17.21",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-hot-loader": "4.13.0"
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   }
 }
