@@ -1,6 +1,5 @@
 // main application wrapper component
 import * as React from 'react'
-import { hot } from 'react-hot-loader/root'
 import cx from 'classnames'
 
 import { DefinitionRoute } from '../../definitions'
@@ -49,8 +48,6 @@ export function AppComponent(props: DefinitionRouteRenderProps): JSX.Element {
   )
 }
 
-export function AppWithRoute(): JSX.Element {
+export function App(): JSX.Element {
   return <DefinitionRoute render={props => <AppComponent {...props} />} />
 }
-
-export const App = hot(AppWithRoute)
