@@ -2,7 +2,6 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 import { I18nextProvider } from 'react-i18next'
-import { AppContainer } from 'react-hot-loader'
 
 import { configureStore } from './configureStore'
 import { App } from './components/App'
@@ -17,24 +16,18 @@ initialize(store)
 // initialize analytics
 initializeMixpanel(store.getState())
 
+const container = document.getElementById('root')
+
+const root = ReactDOM.createRoot(container)
+
 const render = (Component: any): void => {
-  ReactDOM.render(
+  root.render(
     <Provider store={store}>
-      <AppContainer>
-        <I18nextProvider i18n={i18n as any}>
-          <Component />
-        </I18nextProvider>
-      </AppContainer>
-    </Provider>,
-    document.getElementById('root')
+      <I18nextProvider i18n={i18n as any}>
+        <Component />
+      </I18nextProvider>
+    </Provider>
   )
 }
 
 render(App)
-
-// Hot Module Replacement API
-if (module.hot) {
-  module.hot.accept('./components/App', () => {
-    render(App)
-  })
-}
